backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
httpx>=0.25.0
beautifulsoup4>=4.12.0
playwright>=1.40.0
python-dotenv>=1.0.0
pydantic>=2.5.0
aiosqlite>=0.19.0
lxml>=4.9.0
